+++ merk/src/merk/mod.rs
@@ -34,86 +34,43 @@ pub(crate) mod defaults;
 
 pub mod options;
 
+pub mod apply;
+pub mod clear;
+pub mod committer;
+pub mod get;
+pub mod open;
+pub mod prove;
 pub mod restore;
+pub mod source;
 
 use std::{
     cell::Cell,
-    cmp::Ordering,
     collections::{BTreeSet, LinkedList},
     fmt,
 };
 
+use committer::MerkCommitter;
 use grovedb_costs::{
     cost_return_on_error, cost_return_on_error_default, cost_return_on_error_no_add,
     storage_cost::{
-        key_value_cost::KeyValueStorageCost,
-        removal::{StorageRemovedBytes, StorageRemovedBytes::BasicStorageRemoval},
-        StorageCost,
+        key_value_cost::KeyValueStorageCost, removal::StorageRemovedBytes, StorageCost,
     },
     ChildrenSizesWithValue, CostContext, CostResult, CostsExt, FeatureSumLength, OperationCost,
 };
 use grovedb_storage::{self, Batch, RawIterator, StorageContext};
+use source::MerkSource;
 
 use crate::{
     error::Error,
-    merk::{
-        defaults::{MAX_UPDATE_VALUE_BASED_ON_COSTS_TIMES, ROOT_KEY_KEY},
-        options::MerkOptions,
-    },
-    proofs::{encode_into, query::query_item::QueryItem, Op as ProofOp, Query},
+    merk::{defaults::ROOT_KEY_KEY, options::MerkOptions},
+    proofs::{query::query_item::QueryItem, Op as ProofOp, Query},
     tree::{
-        kv::{ValueDefinedCostType, KV},
-        AuxMerkBatch, Commit, CryptoHash, Fetch, Link, MerkBatch, Op, RefWalker, Tree, Walker,
+        kv::ValueDefinedCostType, AuxMerkBatch, Commit, CryptoHash, Fetch, Op, RefWalker, Tree,
         NULL_HASH,
     },
     Error::{CostsError, EdError, StorageError},
     MerkType::{BaseMerk, LayeredMerk, StandaloneMerk},
-    TreeFeatureType,
 };
-
-type Proof = (LinkedList<ProofOp>, Option<u16>, Option<u16>);
-
-/// Proof construction result
-pub struct ProofConstructionResult {
-    /// Proof
-    pub proof: Vec<u8>,
-    /// Limit
-    pub limit: Option<u16>,
-    /// Offset
-    pub offset: Option<u16>,
-}
-
-impl ProofConstructionResult {
-    /// New ProofConstructionResult
-    pub fn new(proof: Vec<u8>, limit: Option<u16>, offset: Option<u16>) -> Self {
-        Self {
-            proof,
-            limit,
-            offset,
-        }
-    }
-}
-
-/// Proof without encoding result
-pub struct ProofWithoutEncodingResult {
-    /// Proof
-    pub proof: LinkedList<ProofOp>,
-    /// Limit
-    pub limit: Option<u16>,
-    /// Offset
-    pub offset: Option<u16>,
-}
-
-impl ProofWithoutEncodingResult {
-    /// New ProofWithoutEncodingResult
-    pub fn new(proof: LinkedList<ProofOp>, limit: Option<u16>, offset: Option<u16>) -> Self {
-        Self {
-            proof,
-            limit,
-            offset,
-        }
-    }
-}
 
 /// Key update types
 pub struct KeyUpdates {
@@ -303,291 +260,6 @@ impl<'db, S> Merk<S>
 where
     S: StorageContext<'db>,
 {
-    /// Open empty tree
-    pub fn open_empty(storage: S, merk_type: MerkType, is_sum_tree: bool) -> Self {
-        Self {
-            tree: Cell::new(None),
-            root_tree_key: Cell::new(None),
-            storage,
-            merk_type,
-            is_sum_tree,
-        }
-    }
-
-    /// Open standalone tree
-    pub fn open_standalone(storage: S, is_sum_tree: bool) -> CostResult<Self, Error> {
-        let mut merk = Self {
-            tree: Cell::new(None),
-            root_tree_key: Cell::new(None),
-            storage,
-            merk_type: StandaloneMerk,
-            is_sum_tree,
-        };
-
-        merk.load_base_root().map_ok(|_| merk)
-    }
-
-    /// Open base tree
-    pub fn open_base(storage: S, is_sum_tree: bool) -> CostResult<Self, Error> {
-        let mut merk = Self {
-            tree: Cell::new(None),
-            root_tree_key: Cell::new(None),
-            storage,
-            merk_type: BaseMerk,
-            is_sum_tree,
-        };
-
-        merk.load_base_root().map_ok(|_| merk)
-    }
-
-    /// Open layered tree with root key
-    pub fn open_layered_with_root_key(
-        storage: S,
-        root_key: Option<Vec<u8>>,
-        is_sum_tree: bool,
-    ) -> CostResult<Self, Error> {
-        let mut merk = Self {
-            tree: Cell::new(None),
-            root_tree_key: Cell::new(root_key),
-            storage,
-            merk_type: LayeredMerk,
-            is_sum_tree,
-        };
-
-        merk.load_root().map_ok(|_| merk)
-    }
-
-    /// Deletes tree data
-    pub fn clear(&mut self) -> CostResult<(), Error> {
-        let mut cost = OperationCost::default();
-
-        let mut iter = self.storage.raw_iter();
-        iter.seek_to_first().unwrap_add_cost(&mut cost);
-
-        let mut to_delete = self.storage.new_batch();
-        while iter.valid().unwrap_add_cost(&mut cost) {
-            if let Some(key) = iter.key().unwrap_add_cost(&mut cost) {
-                // todo: deal with cost reimbursement
-                to_delete.delete(key, None);
-            }
-            iter.next().unwrap_add_cost(&mut cost);
-        }
-        cost_return_on_error!(
-            &mut cost,
-            self.storage.commit_batch(to_delete).map_err(StorageError)
-        );
-        self.tree.set(None);
-        Ok(()).wrap_with_cost(cost)
-    }
-
-    /// Gets an auxiliary value.
-    pub fn get_aux(&self, key: &[u8]) -> CostResult<Option<Vec<u8>>, Error> {
-        self.storage.get_aux(key).map_err(StorageError)
-    }
-
-    /// Returns if the value at the given key exists
-    ///
-    /// Note that this is essentially the same as a normal RocksDB `get`, so
-    /// should be a fast operation and has almost no tree overhead.
-    pub fn exists(&self, key: &[u8]) -> CostResult<bool, Error> {
-        self.has_node_direct(key)
-    }
-
-    /// Returns if the value at the given key exists
-    ///
-    /// Note that this is essentially the same as a normal RocksDB `get`, so
-    /// should be a fast operation and has almost no tree overhead.
-    /// Contrary to a simple exists, this traverses the tree and can be faster
-    /// if the tree is cached, but slower if it is not
-    pub fn exists_by_traversing_tree(&self, key: &[u8]) -> CostResult<bool, Error> {
-        self.has_node(key)
-    }
-
-    /// Gets a value for the given key. If the key is not found, `None` is
-    /// returned.
-    ///
-    /// Note that this is essentially the same as a normal RocksDB `get`, so
-    /// should be a fast operation and has almost no tree overhead.
-    pub fn get(&self, key: &[u8], allow_cache: bool) -> CostResult<Option<Vec<u8>>, Error> {
-        if allow_cache {
-            self.get_node_fn(key, |node| {
-                node.value_as_slice()
-                    .to_vec()
-                    .wrap_with_cost(Default::default())
-            })
-        } else {
-            self.get_node_direct_fn(key, |node| {
-                node.value_as_slice()
-                    .to_vec()
-                    .wrap_with_cost(Default::default())
-            })
-        }
-    }
-
-    /// Returns the feature type for the node at the given key.
-    pub fn get_feature_type(
-        &self,
-        key: &[u8],
-        allow_cache: bool,
-    ) -> CostResult<Option<TreeFeatureType>, Error> {
-        if allow_cache {
-            self.get_node_fn(key, |node| {
-                node.feature_type().wrap_with_cost(Default::default())
-            })
-        } else {
-            self.get_node_direct_fn(key, |node| {
-                node.feature_type().wrap_with_cost(Default::default())
-            })
-        }
-    }
-
-    /// Gets a hash of a node by a given key, `None` is returned in case
-    /// when node not found by the key.
-    pub fn get_hash(&self, key: &[u8], allow_cache: bool) -> CostResult<Option<CryptoHash>, Error> {
-        if allow_cache {
-            self.get_node_fn(key, |node| node.hash())
-        } else {
-            self.get_node_direct_fn(key, |node| node.hash())
-        }
-    }
-
-    /// Gets the value hash of a node by a given key, `None` is returned in case
-    /// when node not found by the key.
-    pub fn get_value_hash(
-        &self,
-        key: &[u8],
-        allow_cache: bool,
-    ) -> CostResult<Option<CryptoHash>, Error> {
-        if allow_cache {
-            self.get_node_fn(key, |node| {
-                (*node.value_hash()).wrap_with_cost(OperationCost::default())
-            })
-        } else {
-            self.get_node_direct_fn(key, |node| {
-                (*node.value_hash()).wrap_with_cost(OperationCost::default())
-            })
-        }
-    }
-
-    /// Gets a hash of a node by a given key, `None` is returned in case
-    /// when node not found by the key.
-    pub fn get_kv_hash(
-        &self,
-        key: &[u8],
-        allow_cache: bool,
-    ) -> CostResult<Option<CryptoHash>, Error> {
-        if allow_cache {
-            self.get_node_fn(key, |node| {
-                (*node.inner.kv.hash()).wrap_with_cost(OperationCost::default())
-            })
-        } else {
-            self.get_node_direct_fn(key, |node| {
-                (*node.inner.kv.hash()).wrap_with_cost(OperationCost::default())
-            })
-        }
-    }
-
-    /// Gets the value and value hash of a node by a given key, `None` is
-    /// returned in case when node not found by the key.
-    pub fn get_value_and_value_hash(
-        &self,
-        key: &[u8],
-        allow_cache: bool,
-    ) -> CostResult<Option<(Vec<u8>, CryptoHash)>, Error> {
-        if allow_cache {
-            self.get_node_fn(key, |node| {
-                (node.value_as_slice().to_vec(), *node.value_hash())
-                    .wrap_with_cost(OperationCost::default())
-            })
-        } else {
-            self.get_node_direct_fn(key, |node| {
-                (node.value_as_slice().to_vec(), *node.value_hash())
-                    .wrap_with_cost(OperationCost::default())
-            })
-        }
-    }
-
-    /// See if a node's field exists
-    fn has_node_direct(&self, key: &[u8]) -> CostResult<bool, Error> {
-        Tree::get(&self.storage, key).map_ok(|x| x.is_some())
-    }
-
-    /// See if a node's field exists
-    fn has_node(&self, key: &[u8]) -> CostResult<bool, Error> {
-        self.use_tree(move |maybe_tree| {
-            let mut cursor = match maybe_tree {
-                None => return Ok(false).wrap_with_cost(Default::default()), // empty tree
-                Some(tree) => tree,
-            };
-
-            loop {
-                if key == cursor.key() {
-                    return Ok(true).wrap_with_cost(OperationCost::default());
-                }
-
-                let left = key < cursor.key();
-                let link = match cursor.link(left) {
-                    None => return Ok(false).wrap_with_cost(Default::default()), // not found
-                    Some(link) => link,
-                };
-
-                let maybe_child = link.tree();
-                match maybe_child {
-                    None => {
-                        // fetch from RocksDB
-                        break self.has_node_direct(key);
-                    }
-                    Some(child) => cursor = child, // traverse to child
-                }
-            }
-        })
-    }
-
-    /// Generic way to get a node's field
-    fn get_node_direct_fn<T, F>(&self, key: &[u8], f: F) -> CostResult<Option<T>, Error>
-    where
-        F: FnOnce(&Tree) -> CostContext<T>,
-    {
-        Tree::get(&self.storage, key).flat_map_ok(|maybe_node| {
-            let mut cost = OperationCost::default();
-            Ok(maybe_node.map(|node| f(&node).unwrap_add_cost(&mut cost))).wrap_with_cost(cost)
-        })
-    }
-
-    /// Generic way to get a node's field
-    fn get_node_fn<T, F>(&self, key: &[u8], f: F) -> CostResult<Option<T>, Error>
-    where
-        F: FnOnce(&Tree) -> CostContext<T>,
-    {
-        self.use_tree(move |maybe_tree| {
-            let mut cursor = match maybe_tree {
-                None => return Ok(None).wrap_with_cost(Default::default()), // empty tree
-                Some(tree) => tree,
-            };
-
-            loop {
-                if key == cursor.key() {
-                    return f(cursor).map(|x| Ok(Some(x)));
-                }
-
-                let left = key < cursor.key();
-                let link = match cursor.link(left) {
-                    None => return Ok(None).wrap_with_cost(Default::default()), // not found
-                    Some(link) => link,
-                };
-
-                let maybe_child = link.tree();
-                match maybe_child {
-                    None => {
-                        // fetch from RocksDB
-                        break self.get_node_direct_fn(key, f);
-                    }
-                    Some(child) => cursor = child, // traverse to child
-                }
-            }
-        })
-    }
-
     /// Returns the root hash of the tree (a digest for the entire store which
     /// proofs can be checked against). If the tree is empty, returns the null
     /// hash (zero-filled).
@@ -622,384 +294,6 @@ where
                 tree.hash()
                     .map(|hash| Ok((hash, Some(tree.key().to_vec()), sum)))
             }
-        })
-    }
-
-    /// Applies a batch of operations (puts and deletes) to the tree.
-    ///
-    /// This will fail if the keys in `batch` are not sorted and unique. This
-    /// check creates some overhead, so if you are sure your batch is sorted and
-    /// unique you can use the unsafe `apply_unchecked` for a small performance
-    /// gain.
-    ///
-    /// # Example
-    /// ```
-    /// # let mut store = grovedb_merk::test_utils::TempMerk::new();
-    /// # store.apply::<_, Vec<_>>(&[(vec![4,5,6], Op::Put(vec![0], BasicMerk))], &[], None)
-    ///         .unwrap().expect("");
-    ///
-    /// use grovedb_merk::Op;
-    /// use grovedb_merk::TreeFeatureType::BasicMerk;
-    ///
-    /// let batch = &[
-    ///     // puts value [4,5,6] to key[1,2,3]
-    ///     (vec![1, 2, 3], Op::Put(vec![4, 5, 6], BasicMerk)),
-    ///     // deletes key [4,5,6]
-    ///     (vec![4, 5, 6], Op::Delete),
-    /// ];
-    /// store.apply::<_, Vec<_>>(batch, &[], None).unwrap().expect("");
-    /// ```
-    pub fn apply<KB, KA>(
-        &mut self,
-        batch: &MerkBatch<KB>,
-        aux: &AuxMerkBatch<KA>,
-        options: Option<MerkOptions>,
-    ) -> CostResult<(), Error>
-    where
-        KB: AsRef<[u8]>,
-        KA: AsRef<[u8]>,
-    {
-        let use_sum_nodes = self.is_sum_tree;
-        self.apply_with_costs_just_in_time_value_update(
-            batch,
-            aux,
-            options,
-            &|key, value| {
-                Ok(KV::layered_value_byte_cost_size_for_key_and_value_lengths(
-                    key.len() as u32,
-                    value.len() as u32,
-                    use_sum_nodes,
-                ))
-            },
-            &mut |_costs, _old_value, _value| Ok((false, None)),
-            &mut |_a, key_bytes_to_remove, value_bytes_to_remove| {
-                Ok((
-                    BasicStorageRemoval(key_bytes_to_remove),
-                    BasicStorageRemoval(value_bytes_to_remove),
-                ))
-            },
-        )
-    }
-
-    /// Applies a batch of operations (puts and deletes) to the tree.
-    ///
-    /// This will fail if the keys in `batch` are not sorted and unique. This
-    /// check creates some overhead, so if you are sure your batch is sorted and
-    /// unique you can use the unsafe `apply_unchecked` for a small performance
-    /// gain.
-    ///
-    /// # Example
-    /// ```
-    /// # let mut store = grovedb_merk::test_utils::TempMerk::new();
-    /// # store.apply::<_, Vec<_>>(&[(vec![4,5,6], Op::Put(vec![0], BasicMerk))], &[], None)
-    ///         .unwrap().expect("");
-    ///
-    /// use grovedb_merk::Op;
-    /// use grovedb_merk::TreeFeatureType::BasicMerk;
-    ///
-    /// let batch = &[
-    ///     // puts value [4,5,6] to key[1,2,3]
-    ///     (vec![1, 2, 3], Op::Put(vec![4, 5, 6], BasicMerk)),
-    ///     // deletes key [4,5,6]
-    ///     (vec![4, 5, 6], Op::Delete),
-    /// ];
-    /// store.apply::<_, Vec<_>>(batch, &[], None).unwrap().expect("");
-    /// ```
-    pub fn apply_with_specialized_costs<KB, KA>(
-        &mut self,
-        batch: &MerkBatch<KB>,
-        aux: &AuxMerkBatch<KA>,
-        options: Option<MerkOptions>,
-        old_specialized_cost: &impl Fn(&Vec<u8>, &Vec<u8>) -> Result<u32, Error>,
-    ) -> CostResult<(), Error>
-    where
-        KB: AsRef<[u8]>,
-        KA: AsRef<[u8]>,
-    {
-        self.apply_with_costs_just_in_time_value_update(
-            batch,
-            aux,
-            options,
-            old_specialized_cost,
-            &mut |_costs, _old_value, _value| Ok((false, None)),
-            &mut |_a, key_bytes_to_remove, value_bytes_to_remove| {
-                Ok((
-                    BasicStorageRemoval(key_bytes_to_remove),
-                    BasicStorageRemoval(value_bytes_to_remove),
-                ))
-            },
-        )
-    }
-
-    /// Applies a batch of operations (puts and deletes) to the tree with the
-    /// ability to update values based on costs.
-    ///
-    /// This will fail if the keys in `batch` are not sorted and unique. This
-    /// check creates some overhead, so if you are sure your batch is sorted and
-    /// unique you can use the unsafe `apply_unchecked` for a small performance
-    /// gain.
-    ///
-    /// # Example
-    /// ```
-    /// # let mut store = grovedb_merk::test_utils::TempMerk::new();
-    /// # store.apply_with_costs_just_in_time_value_update::<_, Vec<_>>(
-    ///     &[(vec![4,5,6], Op::Put(vec![0], BasicMerk))],
-    ///     &[],
-    ///     None,
-    ///     &|k, v| Ok(0),
-    ///     &mut |s, v, o| Ok((false, None)),
-    ///     &mut |s, k, v| Ok((NoStorageRemoval, NoStorageRemoval))
-    /// ).unwrap().expect("");
-    ///
-    /// use grovedb_costs::storage_cost::removal::StorageRemovedBytes::NoStorageRemoval;
-    /// use grovedb_merk::Op;
-    /// use grovedb_merk::TreeFeatureType::BasicMerk;
-    ///
-    /// let batch = &[
-    ///     // puts value [4,5,6] to key[1,2,3]
-    ///     (vec![1, 2, 3], Op::Put(vec![4, 5, 6], BasicMerk)),
-    ///     // deletes key [4,5,6]
-    ///     (vec![4, 5, 6], Op::Delete),
-    /// ];
-    ///
-    /// store.apply_with_costs_just_in_time_value_update::<_, Vec<_>>(
-    ///     batch,
-    ///     &[],
-    ///     None,
-    ///     &|k, v| Ok(0),
-    ///     &mut |s, v, o| Ok((false, None)),
-    ///     &mut |s, k, v| Ok((NoStorageRemoval, NoStorageRemoval))
-    /// ).unwrap().expect("");
-    /// ```
-    pub fn apply_with_costs_just_in_time_value_update<KB, KA>(
-        &mut self,
-        batch: &MerkBatch<KB>,
-        aux: &AuxMerkBatch<KA>,
-        options: Option<MerkOptions>,
-        old_specialized_cost: &impl Fn(&Vec<u8>, &Vec<u8>) -> Result<u32, Error>,
-        update_tree_value_based_on_costs: &mut impl FnMut(
-            &StorageCost,
-            &Vec<u8>,
-            &mut Vec<u8>,
-        ) -> Result<
-            (bool, Option<ValueDefinedCostType>),
-            Error,
-        >,
-        section_removal_bytes: &mut impl FnMut(
-            &Vec<u8>,
-            u32,
-            u32,
-        ) -> Result<
-            (StorageRemovedBytes, StorageRemovedBytes),
-            Error,
-        >,
-    ) -> CostResult<(), Error>
-    where
-        KB: AsRef<[u8]>,
-        KA: AsRef<[u8]>,
-    {
-        // ensure keys in batch are sorted and unique
-        let mut maybe_prev_key: Option<&KB> = None;
-        for (key, ..) in batch.iter() {
-            if let Some(prev_key) = maybe_prev_key {
-                match prev_key.as_ref().cmp(key.as_ref()) {
-                    Ordering::Greater => {
-                        return Err(Error::InvalidInputError("Keys in batch must be sorted"))
-                            .wrap_with_cost(Default::default())
-                    }
-                    Ordering::Equal => {
-                        return Err(Error::InvalidInputError("Keys in batch must be unique"))
-                            .wrap_with_cost(Default::default())
-                    }
-                    _ => (),
-                }
-            }
-            maybe_prev_key = Some(key);
-        }
-
-        self.apply_unchecked(
-            batch,
-            aux,
-            options,
-            old_specialized_cost,
-            update_tree_value_based_on_costs,
-            section_removal_bytes,
-        )
-    }
-
-    /// Applies a batch of operations (puts and deletes) to the tree.
-    ///
-    /// # Safety
-    /// This is unsafe because the keys in `batch` must be sorted and unique -
-    /// if they are not, there will be undefined behavior. For a safe version of
-    /// this method which checks to ensure the batch is sorted and unique, see
-    /// `apply`.
-    ///
-    /// # Example
-    /// ```
-    /// # let mut store = grovedb_merk::test_utils::TempMerk::new();
-    /// # store.apply_with_costs_just_in_time_value_update::<_, Vec<_>>(
-    ///     &[(vec![4,5,6], Op::Put(vec![0], BasicMerk))],
-    ///     &[],
-    ///     None,
-    ///     &|k, v| Ok(0),
-    ///     &mut |s, o, v| Ok((false, None)),
-    ///     &mut |s, k, v| Ok((NoStorageRemoval, NoStorageRemoval))
-    /// ).unwrap().expect("");
-    ///
-    /// use grovedb_costs::storage_cost::removal::StorageRemovedBytes::NoStorageRemoval;
-    /// use grovedb_merk::Op;
-    /// use grovedb_merk::TreeFeatureType::BasicMerk;
-    ///
-    /// let batch = &[
-    ///     // puts value [4,5,6] to key [1,2,3]
-    ///     (vec![1, 2, 3], Op::Put(vec![4, 5, 6], BasicMerk)),
-    ///     // deletes key [4,5,6]
-    ///     (vec![4, 5, 6], Op::Delete),
-    /// ];
-    ///     unsafe { store.apply_unchecked::<_, Vec<_>, _, _, _>(    /// /// ///
-    ///     batch,
-    ///     &[],
-    ///     None,
-    ///     &|k, v| Ok(0),
-    ///     &mut |s, o, v| Ok((false, None)),
-    ///     &mut |s, k, v| Ok((NoStorageRemoval, NoStorageRemoval))
-    /// ).unwrap().expect("");
-    /// }
-    /// ```
-    pub fn apply_unchecked<KB, KA, C, U, R>(
-        &mut self,
-        batch: &MerkBatch<KB>,
-        aux: &AuxMerkBatch<KA>,
-        options: Option<MerkOptions>,
-        old_specialized_cost: &C,
-        update_tree_value_based_on_costs: &mut U,
-        section_removal_bytes: &mut R,
-    ) -> CostResult<(), Error>
-    where
-        KB: AsRef<[u8]>,
-        KA: AsRef<[u8]>,
-        C: Fn(&Vec<u8>, &Vec<u8>) -> Result<u32, Error>,
-        U: FnMut(
-            &StorageCost,
-            &Vec<u8>,
-            &mut Vec<u8>,
-        ) -> Result<(bool, Option<ValueDefinedCostType>), Error>,
-        R: FnMut(&Vec<u8>, u32, u32) -> Result<(StorageRemovedBytes, StorageRemovedBytes), Error>,
-    {
-        let maybe_walker = self
-            .tree
-            .take()
-            .take()
-            .map(|tree| Walker::new(tree, self.source()));
-
-        Walker::apply_to(
-            maybe_walker,
-            batch,
-            self.source(),
-            old_specialized_cost,
-            section_removal_bytes,
-        )
-        .flat_map_ok(|(maybe_tree, key_updates)| {
-            // we set the new root node of the merk tree
-            self.tree.set(maybe_tree);
-            // commit changes to db
-            self.commit(
-                key_updates,
-                aux,
-                options,
-                old_specialized_cost,
-                update_tree_value_based_on_costs,
-                section_removal_bytes,
-            )
-        })
-    }
-
-    /// Creates a Merkle proof for the list of queried keys. For each key in the
-    /// query, if the key is found in the store then the value will be proven to
-    /// be in the tree. For each key in the query that does not exist in the
-    /// tree, its absence will be proven by including boundary keys.
-    ///
-    /// The proof returned is in an encoded format which can be verified with
-    /// `merk::verify`.
-    ///
-    /// This will fail if the keys in `query` are not sorted and unique. This
-    /// check adds some overhead, so if you are sure your batch is sorted and
-    /// unique you can use the unsafe `prove_unchecked` for a small performance
-    /// gain.
-    pub fn prove(
-        &self,
-        query: Query,
-        limit: Option<u16>,
-        offset: Option<u16>,
-    ) -> CostResult<ProofConstructionResult, Error> {
-        let left_to_right = query.left_to_right;
-        self.prove_unchecked(query, limit, offset, left_to_right)
-            .map_ok(|(proof, limit, offset)| {
-                let mut bytes = Vec::with_capacity(128);
-                encode_into(proof.iter(), &mut bytes);
-                ProofConstructionResult::new(bytes, limit, offset)
-            })
-    }
-
-    /// Creates a Merkle proof for the list of queried keys. For each key in the
-    /// query, if the key is found in the store then the value will be proven to
-    /// be in the tree. For each key in the query that does not exist in the
-    /// tree, its absence will be proven by including boundary keys.
-    ///
-    /// The proof returned is in an intermediate format to be later encoded
-    ///
-    /// This will fail if the keys in `query` are not sorted and unique. This
-    /// check adds some overhead, so if you are sure your batch is sorted and
-    /// unique you can use the unsafe `prove_unchecked` for a small performance
-    /// gain.
-    pub fn prove_without_encoding(
-        &self,
-        query: Query,
-        limit: Option<u16>,
-        offset: Option<u16>,
-    ) -> CostResult<ProofWithoutEncodingResult, Error> {
-        let left_to_right = query.left_to_right;
-        self.prove_unchecked(query, limit, offset, left_to_right)
-            .map_ok(|(proof, limit, offset)| ProofWithoutEncodingResult::new(proof, limit, offset))
-    }
-
-    /// Creates a Merkle proof for the list of queried keys. For each key in
-    /// the query, if the key is found in the store then the value will be
-    /// proven to be in the tree. For each key in the query that does not
-    /// exist in the tree, its absence will be proven by including
-    /// boundary keys.
-    /// The proof returned is in an encoded format which can be verified with
-    /// `merk::verify`.
-    ///
-    /// This is unsafe because the keys in `query` must be sorted and unique -
-    /// if they are not, there will be undefined behavior. For a safe version
-    /// of this method which checks to ensure the batch is sorted and
-    /// unique, see `prove`.
-    pub fn prove_unchecked<Q, I>(
-        &self,
-        query: I,
-        limit: Option<u16>,
-        offset: Option<u16>,
-        left_to_right: bool,
-    ) -> CostResult<Proof, Error>
-    where
-        Q: Into<QueryItem>,
-        I: IntoIterator<Item = Q>,
-    {
-        let query_vec: Vec<QueryItem> = query.into_iter().map(Into::into).collect();
-
-        self.use_tree_mut(|maybe_tree| {
-            maybe_tree
-                .ok_or(Error::CorruptedCodeExecution(
-                    "Cannot create proof for empty tree",
-                ))
-                .wrap_with_cost(Default::default())
-                .flat_map_ok(|tree| {
-                    let mut ref_walker = RefWalker::new(tree, self.source());
-                    ref_walker.create_proof(query_vec.as_slice(), limit, offset, left_to_right)
-                })
-                .map_ok(|(proof, _, limit, offset, ..)| (proof, limit, offset))
         })
     }
 
@@ -1186,13 +480,6 @@ where
         true.wrap_with_cost(cost)
     }
 
-    fn source(&self) -> MerkSource<S> {
-        MerkSource {
-            storage: &self.storage,
-            is_sum_tree: self.is_sum_tree,
-        }
-    }
-
     /// Use tree
     pub(crate) fn use_tree<T>(&self, f: impl FnOnce(Option<&Tree>) -> T) -> T {
         let tree = self.tree.take();
@@ -1276,173 +563,19 @@ fn fetch_node<'db>(db: &impl StorageContext<'db>, key: &[u8]) -> Result<Option<T
     }
 }
 
-// impl Clone for Merk<S> {
-//     fn clone(&self) -> Self {
-//         let tree_clone = match self.tree.take() {
-//             None => None,
-//             Some(tree) => {
-//                 let clone = tree.clone();
-//                 self.tree.set(Some(tree));
-//                 Some(clone)
-//             }
-//         };
-//         Self {
-//             tree: Cell::new(tree_clone),
-//             storage_cost: self.storage_cost.clone(),
-//         }
-//     }
-// }
-
 // // TODO: get rid of Fetch/source and use GroveDB storage_cost abstraction
-
-#[derive(Debug)]
-pub struct MerkSource<'s, S> {
-    storage: &'s S,
-    is_sum_tree: bool,
-}
-
-impl<'s, S> Clone for MerkSource<'s, S> {
-    fn clone(&self) -> Self {
-        MerkSource {
-            storage: self.storage,
-            is_sum_tree: self.is_sum_tree,
-        }
-    }
-}
-
-impl<'s, 'db, S> Fetch for MerkSource<'s, S>
-where
-    S: StorageContext<'db>,
-{
-    fn fetch(&self, link: &Link) -> CostResult<Tree, Error> {
-        Tree::get(self.storage, link.key())
-            .map_ok(|x| x.ok_or(Error::KeyNotFoundError("Key not found for fetch")))
-            .flatten()
-    }
-}
-
-struct MerkCommitter {
-    /// The batch has a key, maybe a value, with the value bytes, maybe the left
-    /// child size and maybe the right child size, then the
-    /// key_value_storage_cost
-    batch: Vec<BatchValue>,
-    height: u8,
-    levels: u8,
-}
-
-impl MerkCommitter {
-    fn new(height: u8, levels: u8) -> Self {
-        Self {
-            batch: Vec::with_capacity(10000),
-            height,
-            levels,
-        }
-    }
-}
-
-impl Commit for MerkCommitter {
-    fn write(
-        &mut self,
-        tree: &mut Tree,
-        old_specialized_cost: &impl Fn(&Vec<u8>, &Vec<u8>) -> Result<u32, Error>,
-        update_tree_value_based_on_costs: &mut impl FnMut(
-            &StorageCost,
-            &Vec<u8>,
-            &mut Vec<u8>,
-        ) -> Result<
-            (bool, Option<ValueDefinedCostType>),
-            Error,
-        >,
-        section_removal_bytes: &mut impl FnMut(
-            &Vec<u8>,
-            u32,
-            u32,
-        ) -> Result<
-            (StorageRemovedBytes, StorageRemovedBytes),
-            Error,
-        >,
-    ) -> Result<(), Error> {
-        let tree_size = tree.encoding_length();
-        let (mut current_tree_plus_hook_size, mut storage_costs) =
-            tree.kv_with_parent_hook_size_and_storage_cost(old_specialized_cost)?;
-        let mut i = 0;
-
-        if let Some(old_value) = tree.old_value.clone() {
-            // At this point the tree value can be updated based on client requirements
-            // For example to store the costs
-            loop {
-                let (flags_changed, value_defined_cost) = update_tree_value_based_on_costs(
-                    &storage_costs.value_storage_cost,
-                    &old_value,
-                    tree.value_mut_ref(),
-                )?;
-                if !flags_changed {
-                    break;
-                } else {
-                    tree.inner.kv.value_defined_cost = value_defined_cost;
-                    let after_update_tree_plus_hook_size =
-                        tree.value_encoding_length_with_parent_to_child_reference();
-                    if after_update_tree_plus_hook_size == current_tree_plus_hook_size {
-                        break;
-                    }
-                    let new_size_and_storage_costs =
-                        tree.kv_with_parent_hook_size_and_storage_cost(old_specialized_cost)?;
-                    current_tree_plus_hook_size = new_size_and_storage_costs.0;
-                    storage_costs = new_size_and_storage_costs.1;
-                }
-                if i > MAX_UPDATE_VALUE_BASED_ON_COSTS_TIMES {
-                    return Err(Error::CyclicError(
-                        "updated value based on costs too many times",
-                    ));
-                }
-                i += 1;
-            }
-
-            if let BasicStorageRemoval(removed_bytes) =
-                storage_costs.value_storage_cost.removed_bytes
-            {
-                let (_, value_removed_bytes) = section_removal_bytes(&old_value, 0, removed_bytes)?;
-                storage_costs.value_storage_cost.removed_bytes = value_removed_bytes;
-            }
-        }
-
-        // Update old tree size after generating value storage_cost cost
-        tree.old_size_with_parent_to_child_hook = current_tree_plus_hook_size;
-        tree.old_value = Some(tree.value_ref().clone());
-
-        let mut buf = Vec::with_capacity(tree_size);
-        tree.encode_into(&mut buf);
-
-        let left_child_sizes = tree.child_ref_and_sum_size(true);
-        let right_child_sizes = tree.child_ref_and_sum_size(false);
-        self.batch.push((
-            tree.key().to_vec(),
-            tree.feature_type().sum_length(),
-            Some((buf, left_child_sizes, right_child_sizes)),
-            Some(storage_costs),
-        ));
-        Ok(())
-    }
-
-    fn prune(&self, tree: &Tree) -> (bool, bool) {
-        // keep N top levels of tree
-        let prune = (self.height - tree.height()) >= self.levels;
-        (prune, prune)
-    }
-}
 
 #[cfg(test)]
 mod test {
-    use grovedb_costs::OperationCost;
     use grovedb_path::SubtreePath;
     use grovedb_storage::{
-        rocksdb_storage::{test_utils::TempStorage, PrefixedRocksDbStorageContext, RocksDbStorage},
+        rocksdb_storage::{PrefixedRocksDbStorageContext, RocksDbStorage},
         RawIterator, Storage, StorageBatch, StorageContext,
     };
     use tempfile::TempDir;
 
-    use super::{Merk, MerkSource, RefWalker};
-    use crate::{test_utils::*, Op, TreeFeatureType::BasicMerk};
+    use super::{Merk, RefWalker};
+    use crate::{merk::source::MerkSource, test_utils::*, Op, TreeFeatureType::BasicMerk};
 
     // TODO: Close and then reopen test
 
@@ -1451,97 +584,6 @@ mod test {
             let tree = maybe_tree.expect("expected tree");
             assert_tree_invariants(tree);
         })
-    }
-
-    #[test]
-    fn test_reopen_root_hash() {
-        let tmp_dir = TempDir::new().expect("cannot open tempdir");
-        let storage = RocksDbStorage::default_rocksdb_with_path(tmp_dir.path())
-            .expect("cannot open rocksdb storage");
-        let test_prefix = [b"ayy"];
-
-        let batch = StorageBatch::new();
-        let mut merk = Merk::open_base(
-            storage
-                .get_storage_context(SubtreePath::from(test_prefix.as_ref()), Some(&batch))
-                .unwrap(),
-            false,
-        )
-        .unwrap()
-        .unwrap();
-
-        merk.apply::<_, Vec<_>>(
-            &[(vec![1, 2, 3], Op::Put(vec![4, 5, 6], BasicMerk))],
-            &[],
-            None,
-        )
-        .unwrap()
-        .expect("apply failed");
-
-        let root_hash = merk.root_hash();
-
-        storage
-            .commit_multi_context_batch(batch, None)
-            .unwrap()
-            .expect("cannot commit batch");
-
-        let merk = Merk::open_base(
-            storage
-                .get_storage_context(SubtreePath::from(test_prefix.as_ref()), None)
-                .unwrap(),
-            false,
-        )
-        .unwrap()
-        .unwrap();
-        assert_eq!(merk.root_hash(), root_hash);
-    }
-
-    #[test]
-    fn test_open_fee() {
-        let storage = TempStorage::new();
-        let batch = StorageBatch::new();
-
-        let merk_fee_context = Merk::open_base(
-            storage
-                .get_storage_context(SubtreePath::empty(), Some(&batch))
-                .unwrap(),
-            false,
-        );
-
-        // Opening not existing merk should cost only root key seek (except context
-        // creation)
-        assert!(matches!(
-            merk_fee_context.cost(),
-            OperationCost { seek_count: 1, .. }
-        ));
-
-        let mut merk = merk_fee_context.unwrap().unwrap();
-        merk.apply::<_, Vec<_>>(
-            &[(vec![1, 2, 3], Op::Put(vec![4, 5, 6], BasicMerk))],
-            &[],
-            None,
-        )
-        .unwrap()
-        .expect("apply failed");
-
-        storage
-            .commit_multi_context_batch(batch, None)
-            .unwrap()
-            .expect("cannot commit batch");
-
-        let merk_fee_context = Merk::open_base(
-            storage
-                .get_storage_context(SubtreePath::empty(), None)
-                .unwrap(),
-            false,
-        );
-
-        // Opening existing merk should cost two seeks. (except context creation)
-        assert!(matches!(
-            merk_fee_context.cost(),
-            OperationCost { seek_count: 2, .. }
-        ));
-        assert!(merk_fee_context.cost().storage_loaded_bytes > 0);
     }
 
     #[test]
@@ -1579,29 +621,6 @@ mod test {
             .unwrap()
             .expect("apply failed");
         assert_invariants(&merk);
-    }
-
-    #[test]
-    fn test_has_node_with_empty_tree() {
-        let mut merk = TempMerk::new();
-
-        let key = b"something";
-
-        let result = merk.has_node(key).unwrap().unwrap();
-
-        assert!(!result);
-
-        let batch_entry = (key, Op::Put(vec![123; 60], BasicMerk));
-
-        let batch = vec![batch_entry];
-
-        merk.apply::<_, Vec<_>>(&batch, &[], None)
-            .unwrap()
-            .expect("should ...");
-
-        let result = merk.has_node(key).unwrap().unwrap();
-
-        assert!(result);
     }
 
     #[test]
